requirements.txt
# backend requirements
django==4.2.11
djangorestframework==3.15.1
drf-spectacular==0.27.2
# bot requirements
aiogram==3.4.1
python-dotenv==1.0.1
httpx==0.27.0

black==24.3.0
isort==5.13.2
autoflake==2.3.11
